% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsr.R
\name{checkState}
\alias{checkState}
\title{A class to represent the Swedish national school register (NSR).}
\usage{
checkState(object)
}
\description{
A class to represent the Swedish national school register (NSR).
}
\section{Slots}{

\describe{
\item{\code{subjects}}{The list of subjects available in the NSR.}
}}

