% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educationalattainment.R
\name{educationalAttainment}
\alias{educationalAttainment}
\title{Calculate educational attainment level.}
\usage{
educationalAttainment(object, attainments, missing)
}
\arguments{
\item{object}{A vector of SUN2000NIVA codes}

\item{attainments}{A named list of educational attainments to code, eg list(compulsory='mycode1', secondary='mycode2', tertiary='mycode3')}

\item{missing}{The character code for a missing observation, defaults to '999' and are treated as other missing values in data.}
}
\description{
Calculate educational attainment level according to Statistics Swedens classification of education.
}
\details{
This function takes educational attainment codes according to Statistics Swedens classification
(mostly named SUN2000NIVA in datasets) and recodes
them into three levels where compulsory is equivalient of at most 9 years of education, secondary 9 to 12 years, and
tertiary greater than 12 years. Warnings are generated if data contains codes that are unknown according to the offical
documentation from statistics Sweden.
}
\examples{
# 2 with compulsory education, 1 with secondary, 1 with tertiary, 2 missing
sunCodes <- c('001','200','312', '415','999', NA)
newCodes <- educationalAttainment(sunCodes, list(compulsory=1, secondary=2, tertiary=3))
codes
# [1] "1", "1", "2", "3", NA, NA
}
