% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsr.R
\name{GPA}
\alias{GPA}
\alias{GPA,NSR-method}
\title{Calculate grade-point average (GPA)}
\usage{
GPA(object, subjects = F)

\S4method{GPA}{NSR}(object, subjects = F)
}
\arguments{
\item{object}{An object of class NSR.}

\item{subjects}{A list of subjects to include in the GPA}
}
\description{
Calculate grade-point average (GPA)
}
