% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGrade.R
\name{scoreGrade}
\alias{scoreGrade}
\title{Convert alphabetical grade to numerical}
\usage{
scoreGrade(grade, year)
}
\arguments{
\item{grade}{A vector of grades (e.g., A, E, VG, F)}

\item{year}{A vector of graduation years for the corrsponding grade.}
}
\description{
This function takes vectors of year and raw grades in a subject
and converts it into a number between 0 and 20 according to the
official scoring table according to Skolverket.
}
