% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsr.R
\name{setSubjectNames}
\alias{setSubjectNames}
\title{Set subject names}
\usage{
setSubjectNames(object, subjectNames)
}
\arguments{
\item{subjectNames}{A named list of subjects.}
}
\description{
Set subject names
}
