% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsr.R, R/scoreGrade.R
\name{scoreSubjects}
\alias{scoreSubjects}
\alias{scoreSubjects,NSR,vector-method}
\alias{scoreGrade}
\title{Convert alphabetical grades to numeric}
\usage{
scoreSubjects(object, year, ...)

\S4method{scoreSubjects}{NSR,vector}(object, year, ...)

scoreGrade(grade, year)
}
\arguments{
\item{object}{An object of class NSR}

\item{year}{A vector of graduation years for the corrsponding grade.}

\item{...}{Grades to score.}

\item{grade}{A vector of grades (e.g., A, E, VG, F)}
}
\description{
This function takes vectors of year and raw grades in a subject
and converts it into a number between 0 and 20 according to the
official scoring table according to Skolverket.
}
